FDG-8450 refactoring, added title conditional
